style: use components instead of a hook that returns components
